perf(schema/postgres):  Remove fmt.Sprintf call in SetSearchPath
